refactor: close_recruit/ ディレクトリを作成して auto_close, close_event, close_notify_event, close_command を移動
