+++ src/app/feat-recruit/create_recruit/event_recruit.ts
@@ -5,9 +5,9 @@ import { UniqueRoleService } from '../../../db/unique_role_service';
 import { getEventData, EventMatchInfo } from '../../common/apis/splatoon3.ink/splatoon3_ink';
 import { assertExistCheck, exists, sleep } from '../../common/others';
 import { RoleKeySet } from '../../constant/role_key';
-import { recruitAutoClose } from '../auto_close';
 import { recruitEventCanvas, ruleEventCanvas } from '../canvases/event_canvas';
 import { RecruitOpCode } from '../canvases/regenerate_canvas';
+import { recruitAutoClose } from '../close_recruit/auto_close';
 import { arrangeCommandRecruitData } from './common/arrange_command_data';
 import { arrangeModalRecruitData } from './common/arrange_modal_data';
 import { registerRecruitData } from './common/register_recruit_data';
