+++ src/app/feat-recruit/create_recruit/regular_recruit.ts
@@ -5,9 +5,9 @@ import { UniqueRoleService } from '../../../db/unique_role_service';
 import { getRegularData, MatchInfo } from '../../common/apis/splatoon3.ink/splatoon3_ink';
 import { assertExistCheck, exists, sleep } from '../../common/others';
 import { RoleKeySet } from '../../constant/role_key';
-import { recruitAutoClose } from '../auto_close';
 import { RecruitOpCode } from '../canvases/regenerate_canvas';
 import { recruitRegularCanvas, ruleRegularCanvas } from '../canvases/regular_canvas';
+import { recruitAutoClose } from '../close_recruit/auto_close';
 import { arrangeCommandRecruitData } from './common/arrange_command_data';
 import { arrangeModalRecruitData } from './common/arrange_modal_data';
 import { registerRecruitData } from './common/register_recruit_data';
